refactor: loading 컴포넌트를 공통 컴포넌트 폴더로 이동
