+++ src/components/loading/Loading.jsx
@@ -1,4 +1,4 @@
-import StyledLoading from "@/pages/loading/Loading.styled";
+import StyledLoading from "@/components/loading/Loading.styled";
 import BaedaliImg from "@/assets/images/baedali.png";
 
 function Loading() {
