+++ src/components/loading/Loading.styled.js
